Progressive part is working
Fixed the network that now progresses in both resolution and details from Fourier frequencies, once it reaches the target resolution. There are still some things to check, first of all the normalization of the input images. One possible variation that might, counterintuitively, help the network is the zoom-in in the central region ([256:728, 256:728]) since more differences between images emerge.